--- src/features/profile/pages/AfterChangeEmailValidationBuffer/index.ts
@@ -1,1 +1,0 @@
-export * from './AfterChangeEmailValidationBuffer'
